feat: refactor: `MasterFood` inherits a base class `NutritionalContent`
Doing this so that a Food Diary Entry (coming soon) can inherit `NutritionalContent` as well, and act as an aggregate: i.e. number of servings times quantity of some nutritional value per serving equals the total amount of that nutritional value for the food diary entry.  Hopefully this is a good idea...stay tuned! ;-)

Also, made `DeactivatedDateUtc` a `TIMESTAMP` (not a `TIMESTAMPTZ`)